Lab 01
Tried to reach you by email but never got a response. My house was out of power and internet because some AT&T guy accidentally cut our line, so I was unable to submit the lab. I also still can't get pytest or pip to install, so I will need to come to office hours to figure that out